+++ R/zzz.R
@@ -44,5 +44,5 @@ utils::globalVariables(c(
   "q_to",      # plt_hte_rate:      end of a GATES group's share
   "cate_mean", # plt_hte_rate:      mean forest CATE of a GATES group
   # hte-select.R (reuses variable above)
-  "score_sd", "n_vars", "value", "statistic"
+  "score_sd", "n_vars", "value", "statistic", "position", "curve_x"
 ))